Read ServerMapDataPacket data as unsigned bytes.

+++ src/main/java/org/spacehq/mc/protocol/data/game/values/world/map/MapColumnUpdate.java
@@ -2,35 +2,34 @@ package org.spacehq.mc.protocol.data.game.values.world.map;
 
 public class MapColumnUpdate implements MapData {
 
-	private byte x;
-	private byte y;
-	private byte height;
+	private int x;
+	private int y;
+	private int height;
 	private byte colors[];
 
 	/**
 	 * Creates a new map column update instance.
-	 *
-	 * @param x             X of the map column.
-	 * @param y             Y of the data's update range.
-	 * @param height        Height of the data's update range.
-	 * @param fullMapColors The full array of map color data, arranged in order of ascending Y value relative to the given Y value.
+	 * @param x X of the map column.
+	 * @param y Y of the data's update range.
+	 * @param height Height of the data's update range.
+	 * @param colors The array of map color data, arranged in order of ascending Y value relative to the given Y value.
 	 */
 	public MapColumnUpdate(int x, int y, int height, byte colors[]) {
-		this.x = (byte) x;
-		this.y = (byte) y;
-		this.height = (byte) height;
+		this.x = x;
+		this.y = y;
+		this.height = height;
 		this.colors = colors;
 	}
 
-	public byte getX() {
+	public int getX() {
 		return this.x;
 	}
 
-	public byte getY() {
+	public int getY() {
 		return this.y;
 	}
 
-	public byte getHeight() {
+	public int getHeight() {
 		return this.height;
 	}
